Improves language selection and persistence
Uses cookies-next for locale storage and transitions for smoother UI updates when switching languages.
This change replaces the direct manipulation of document.cookie with the cookies-next library for better handling of cookies in a Next.js environment.
It also introduces a transition when changing languages which updates the language and refreshes the router.

The default locale is now set to 'en'.